news-aggregator/requirements.txt
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
newspaper3k==0.2.8
openai==1.12.0
python-dotenv==1.0.1
pyyaml==6.0.1
python-dateutil==2.8.2
html2text==2024.2.26
schedule==1.2.1
supabase>=2.0.0


